little fixes

+++ bakery_store/static/js/app.js
@@ -175,7 +175,7 @@
         }
 
         function formatPrice(price) {
-            return parseInt(price).toLocaleString('en-US', { 
+            return parseFloat(price).toLocaleString('en-US', { 
                 style: 'currency', currency: 'USD' 
             })
         }
